Implement constant folding (includes constant and copy propagation).
The optimization is done "on-the-fly": at each instruction, constant and copy
propagation are performed (using a propagation environment), and then the
expression is constant folded. If this results in a new constant definition,
the propagation environment is updated. If a variable goes out of scope, it is
removed from the propagation environment.